ArtExtract_Soyoung/requirements.txt

torch>=2.0.0
torchvision>=0.15.0
torchmetrics
timm
einops
pytorch-msssim
wandb
opencv-python
transformers>=4.30.0
diffusers>=0.16.0
accelerate
realesrgan
pillow
numpy
matplotlib
scikit-image